Move router and legacy auction to router directory
